Cameraクラスをフォルダ分けとUTF-8化
